refactor: Refactor into more modules
